rename framed_etc.fits to fxxxxx.fits
emldetect has a limit to the length of the string holding the dataset
names. This limit can sometimes be exceeded by the framed_etc.fits
names. Therefore instead of creating framed_etc.fits (with variable
length) we are now using fxxxxx.fits files (with a standard length of
11 characters).  NB before we were able to overwrite framed files from
a previous run; now this is no longer possible so the user should
purge the directory by themselves.